move tree extension under tree-extension
