+++ webui/react/src/omnibar/tree-extension/utils.ts
